docs: mention stepper component
